Pass the str find/rfind/count residual operands as object references
jit_str_find_objs, jit_str_rfind_objs and jit_str_count_objs took the
string, substring and bound objects as i64, so the descr wrappers cast
each argument with `as i64` and the trace recorded CastPtrToInt on the
string. The unroller carried that integer across the loop's Jump, and
once a minor collection moved the string the loop called the search
with the old address (test_email TestMessageAPI.test_get_param_linear_complexity,
SIGSEGV in str_byte_window). The helpers now take PyObjectRef, as
ll_find takes its STR pointers, and the call descr classifies them 'r'.

+++ pyre/pyre-interpreter/src/type_methods.rs
@@ -2028,14 +2028,7 @@ pub fn __majit_wrap_str_descr_find(args: &[PyObjectRef]) -> Result<PyObjectRef, 
     if !bound_is_none_or_exact_int(start) || !bound_is_none_or_exact_int(end) {
         return str_descr_find_slow(args);
     }
-    let n = unsafe {
-        pyre_object::unicodeobject::jit_str_find_objs(
-            args[0] as i64,
-            args[1] as i64,
-            start as i64,
-            end as i64,
-        )
-    };
+    let n = unsafe { pyre_object::unicodeobject::jit_str_find_objs(args[0], args[1], start, end) };
     Ok(w_int_new(n))
 }
 
@@ -2052,14 +2045,7 @@ pub fn __majit_wrap_str_descr_rfind(args: &[PyObjectRef]) -> Result<PyObjectRef,
     if !bound_is_none_or_exact_int(start) || !bound_is_none_or_exact_int(end) {
         return str_descr_rfind_slow(args);
     }
-    let n = unsafe {
-        pyre_object::unicodeobject::jit_str_rfind_objs(
-            args[0] as i64,
-            args[1] as i64,
-            start as i64,
-            end as i64,
-        )
-    };
+    let n = unsafe { pyre_object::unicodeobject::jit_str_rfind_objs(args[0], args[1], start, end) };
     Ok(w_int_new(n))
 }
 
@@ -2076,14 +2062,7 @@ pub fn __majit_wrap_str_descr_count(args: &[PyObjectRef]) -> Result<PyObjectRef,
     if !bound_is_none_or_exact_int(start) || !bound_is_none_or_exact_int(end) {
         return str_descr_count_slow(args);
     }
-    let n = unsafe {
-        pyre_object::unicodeobject::jit_str_count_objs(
-            args[0] as i64,
-            args[1] as i64,
-            start as i64,
-            end as i64,
-        )
-    };
+    let n = unsafe { pyre_object::unicodeobject::jit_str_count_objs(args[0], args[1], start, end) };
     Ok(w_int_new(n))
 }
